Update README.md
The task is to program an island model with variable parameters consisting of an array of locations (for example, 100x20 cells). The locations will be filled with vegetation and animals. Animals can:
eat plants and/or other animals (if there is suitable food in their location),
move (to neighboring locations),
reproduce (if there is a pair in their location),
starve to death or be eaten.

Mandatory part of the task:
Hierarchy of Animals (OOP)
Animal behavior
Multithreading
Statistics on the state of the island on each clock cycle (in the console)

When creating the program, the following were used:
abstract classes
interfaces
annotations
enums
factory pattern
multithreading